# documents/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "GovTech Location Decision Maker API",
        "description": "GovTech Location Decision Maker APIs Information",
        "version": "1.0"
    },
    "servers": [{
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/user/account/{accountId}": {
            "get": {
                "tags": ["User"],
                "summary": "Retrieve user details of a user by account identifier",
                "operationId": "getByAccountId",
                "parameters": [{
                        "name": "accountId",
                        "in": "path",
                        "description": "Account identifier of the user",
                        "required": true,
                        "schema": {
                            "title": "User account id",
                            "type": "string",
                            "description": "Account identifier of the user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return user information.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountResponse"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Request not acceptable according to the Accept-* header.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized. Application/User is not authorized to call this service.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error. Correct request but a problem occurred on the server. Dependent service is not available.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid account identifier supplied.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method other than GET is not supported.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/account/{id}": {
            "get": {
                "tags": ["Account"],
                "summary": "Retrieve account details of a user by account identifier",
                "operationId": "getByAccountId_1",
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "Account identifier of the user",
                        "required": true,
                        "schema": {
                            "title": "User account id",
                            "type": "string",
                            "description": "Account identifier of the user"
                        }
                    }
                ],
                "responses": {
                    "406": {
                        "description": "Request not acceptable according to the Accept-* header.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return user account information.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized. Application/User is not authorized to call this service.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error. Correct request but a problem occurred on the server. Dependent service is not available.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid account identifier supplied.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method other than GET is not supported.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AccountResponse": {
                "title": "User account response",
                "type": "object",
                "properties": {
                    "id": {
                        "title": "User account identifier",
                        "type": "string",
                        "description": "Account identifier of the user"
                    },
                    "userName": {
                        "title": "User account name",
                        "type": "string",
                        "description": "Account name of the user"
                    }
                },
                "description": "Parameters for the user account response"
            },
            "ExceptionResponse": {
                "title": "Exception response",
                "type": "object",
                "properties": {
                    "message": {
                        "title": "Error message",
                        "type": "string",
                        "description": "Exception message"
                    },
                    "time": {
                        "title": "Exception occured time",
                        "type": "string",
                        "description": "Time at which the exceptoin has occured",
                        "format": "date-time"
                    }
                },
                "description": "Parameters for the exception response"
            },
...
        }
    }
}